First shot at a BLAS benchmark. Currently, only GEMV is supported.
The source code calls gemv in a pattern that looks like classical
Gram Schmidt. It may not have the exact memory reuse as iterative
classical gram schmidt.  That is, the vector that is orthogonalized
is not randomized between tests. The code names the phases
"inner Product" and "squash" and times each independently.

Fortran mangling is not handled very well and Cmake should be used.

The Makefiles are a bit complex as well. TODO is to make the build
system more robust. This will pay off in the future as we try to
study more kernels. The complexity of the build system arises from
the various BLAS libraries and the various threading models that each
support.

Currently, you can build the following combos:

OpenBLAS + GNU OpenMP w/GNU compilers
MKL + GNU OpenMP w/GNU compilers
MKL + Intel OpenMP w/intel compilers
Libsci + Cray Threads w/cray compilers

To achieve each build, you have to fiddle with modules, which makes it
a pain to automate.

launcher.sh will then launch the program using aprun. It aggregates
the data into two csv files (details and aggregate). Details has all
timings. Aggregate averages and performs some analysis on this data.

The code depends on MPI as well, but this is solely for communicating
timing information. The timers wrap directly around BLAS calls, with an
MPI_Barrier between different BLAS calls. In the future, it would be nice
to simulate the network traffic as well. E.g., actually do Allreduce's
and truly orthogonalize.

The code should work across multiple physical nodes. But the aggregate
analysis is not prepared for this. The detailed timing data provides sufficient
information that per-node analysis should be doable.