Script for analyizing data using Pandas, NumPy and Matplotlib
This is the current work in progress at a script for handling thread scaling
analysis. The bulk of the scripting is gathering data into a Pandas DataFrame
and then iterating over subsets of the data.

The current data the script it expects should be tabular with columns:
  Label, hostname, rank, np, OMP_NUM_THREADS, OMP_WAIT_POLICY, OMP_PLACES,
  Date, BlasLib, LapackVersion
  m (local), n, time (ns),

The first two lines of attributes seem like a good starting point for what is
needed to characterize threaded app performance.
The final attrbutes (third line) are attributes unique to the benchmark. These
require expert knowledge in the script. That is, for benchmarking gemv,
m variables based on number of processes, while n should be used for grouping data
(Since 'n' indicates the number of vectors of size m that were used).

Currently, the script is aimed at plotting. It does not write files. A next
step is to add a rollup pass, where data is analyzed to identify the best runs.
That is, which combinations of variables produced the shortest runtime.