plottable_df improvements. This function is responsible for aggregating data into a plottable format. The basic idea is to aggregate (sum) values so that we have a single data point at each node count. This poses some problems because it makes little sense to compare aggregate values if the number of aggregates is vastly different between the two.
To handle this, the raw data is always plotted as-is. The percentage of
total time and mpi_factor are additional represenations that attempt to
provide insight into how the timer actually behaved. These two
calculations take the 'num_steps', which is the number of times a
speciifc experiment was repeated. For MueLu constructions this is
typically 20 times, but we also tend to have multiple runs of data. To
compute these metrics, we take the aggregate numsteps and divide the
timer's runtime by it.  This gives some sense of 'average over
experiments', e.g., a sample mean. This is *not* the average of the
Teuchos timers by dividing by the call count. That measure is not well
defined, because timers can be nested, and so the call count can be
large when the bulk of that time was in a very few calls.

Timer ranking should also be improved. Currently, it ranks based off the
single node data. Ranking based off the largest runs was an idea, until
we discovered that repartitioning was exploding. Some type of average
across all runs may be useful.

Flat MPI comparisons.  This is a little tricky because not all timers
always appear. E.g., based on the datasize per node, there may be more
or fewer levels. Also, between OpenMP and Serial the exact timers do not
always appear. Currently, OpenMP with threads=1 is considered flat MPI.