Data aggregation is not right.
The problem shows up when comparing against 'flat mpi' where the number
of samples for each a decomp differs from the number of samples used for
flat mpi. There is no easy way to fix this. Averaging using numsteps is
not good, because you cannot aggregate across multiple experiments.
E.g., collect all SPMV data and attempt analysis. It seems like the only
way to handle this is to use the call counts to average or in some way
weight the observations. This is really messy since the call counts is a
weird number. Timers are allowed to be nested. Every time the timer is
called, call count is incremented. This means that a divide and conquer
algorithm would have a high call count, but the time / callCount is not
an average time all. This complicates finding the expensive timers in a
global sense as well, because looking at data across multiple studies is
difficult to join together.

Basic summation works fine for looking at raw time. These issues come up
when computing the percentage of time, speedup, efficiency, or the
relative cost compared to flat MPI.