Weak and Strong seem ok.
Raw data plotting should probably switch to printing the timer divided by the call count, which is unfortunate, but it is the only way to handle the variety of data I have. In most cases, I do not have the same numsteps for every run, and in some cases we have multiple runs per node that are OK. Aggregating this data becomes a problem, since you cannot merely compare aggregates.

Dividing by call count is not good, because call count is a weird thing. Dividing by it does not represent the average time spent in that timer, because timers can be nested. Division seems OK, since for a specific partitioning with the same solver + problem the algorithm should behave the same.