Work to improve graphics per discussions.
Goals:
the base line (flat MPI, whatever) should be thicker and more obvious
Y axis should be normalized in some way.
figure legend needs to be larger
Font size should be larger for figures that will be shown at smaller
sizes

Done:
spit out the legend as a seperate file

WIP:
Normalizing is complicated, because this means the Y axis is 'shared'
among two different datasets. Currently, the I output shared and 'free'
axes for the current dataset. Adding logic to use an external y-axis
will be hacked for now, but this really entails logic that supports
comparing multiple datasets and from those datums choosing a suitable
normalization.

Bolding is not difficult. Nor is font size