Scripts that will parse logs (any text file) and find teuchos timer tables.
The functionality is encapsulated in a Python package, and is not
coupled to the experiment framework used by the driver.

In the future, the parse will look for the tokens needed to rebuild the
filenames that would have been used by the driver.

The library provies the following:
1) It finds Teuchos TimerMonitor tables
2) It parses single and quad column timings
3) It records the original text as collection of lines
4) It stores the data in a dictionary similar to what PyYAML will
produce by parsing the YAML output emitted by
Teuchos::TimeMonitor::report() when YAML is requested.
5) It can take a YAML file and report the YAML data as a table.

An example driver script is provided.