Added ability for log file parser to recognize experiment names
This allows the routines to generate files that have names as they
would have been used in my driver.

This function is rather fragile, as nested teuchus timer tables
are difficult to handle. Currently, the script will track the last
valid experiment name it found and append it to the metadata of
the most recent teuchos table parsed. Then, it is reset.